scripts/requirements.txt
azure-identity
azure-mgmt-cognitiveservices
